requirements.txt
asgiref==3.11.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==6.0.2
django-allauth==65.15.0
django-axes==8.3.1
django-cleanup==9.0.0
django-environ==0.13.0
Flask==3.1.2
fonttools==4.61.1
gunicorn==25.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.2
packaging==26.0
pillow==12.1.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
Werkzeug==3.1.5
whitenoise==6.12.0
django-q2==1.9.0
django-picklefield==3.4.0
